Fix authentication navigation routing issue
- Created proper root layout (src/app/layout.tsx) with all providers
- Simplified chat layout to just return children
- Simplified auth layout to avoid duplicate providers
- Fixed Next.js App Router layout hierarchy conflicts
- Navigation now works correctly from welcome card and sidebar
- Removed debugging code and cleaned up components

Resolves sign-in/create account button navigation issues.
